==> 2-Manzanas con el codigo OK.ipynb <==
import os
os.chdir("Ruta/a/tu/carpeta")

import pandas as pd
import geopandas as gpd

gdf=gpd.read_file('rawdata/Manzanas_Chimbote_INEI_Oficial.geojson')
gdf.head()

Lo que se ve es que en este gdf que me pasa INEI por correo es que el ID de manzana está compuesto por *UBIGEO + CODZONA + SUFZONA + CODMZNA + SUFMZNA*. 

Sin embargo, cuando uno baja los datos de REDATAM, el ID de manzana está compuesto por *UBIGEO + **CODCCPP** + CODZONA + SUFZONA + CODMZNA + SUFMZNA*. Por tanto, lo que hemos hecho en el cuaderno *1-Recalculando el Codigo de Manzana* ha sido calcular ese nuevo código de manzana para añadirlo al geodataframe y luego poder jalar las variables de REDATAM sin problema.

idmz=pd.read_csv("data/CodigoManzanasOficial.csv")
gdf_nuevo=gdf.drop(["IDMANZANA",              #excluimos la columna que ya no usaremos
                   "fid"],                    #usamos la columna ID como referencia, esta está de más
                   axis=1) 
gdf_nuevo = gpd.GeoDataFrame(gdf_nuevo.merge(idmz, on='ID', how='left'))
gdf_nuevo.head()

list(gdf_nuevo.columns.values)

gdf_nuevo=gdf_nuevo.drop("Unnamed: 0",axis=1)
gdf_nuevo.to_file("data/1-Mzs_Chimbote_Oficial_INEI.geojson",driver='GeoJSON')
