==> 5-Todos los df unidos en el gdf.ipynb <==
import pandas as pd
import geopandas as gpd
import os

os.chdir("Ruta/a/tu/carpeta")

gdf=gpd.read_file("data/1-Mzs_Chimbote_Oficial_INEI.geojson")
refri=pd.read_csv("data/2-Refrigeradora.csv")
hacin=pd.read_csv("data/3-Hacinamiento.csv")

refri.rename(columns={'Cod':'Mz','No_TotalPR':'NoRefri_TotalPR',
                      'No_PR':'NoRefrixMz_PR'},inplace=True)

hacin.rename(columns={'Cod':'Mz'},inplace=True)

gdf_nuevo = gpd.GeoDataFrame(gdf.merge(refri, on='Mz', how='left'))
gdf_nuevo = gpd.GeoDataFrame(gdf_nuevo.merge(hacin, on='Mz', how='left'))

gdf_nuevo.head()

gdf_nuevo.drop(['Unnamed: 0_x', 'Unnamed: 0_y'], axis = 'columns', inplace=True)
list(gdf_nuevo.columns.values)

gdf_nuevo.head()

gdf_nuevo.to_file("data/4-ManzanasConInfo.geojson",driver='GeoJSON')
